requirements.txt
Flask==2.1.3
Werkzeug==2.0.2
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.46
Flask-APScheduler
psycopg2-binary
requests
openai
APScheduler
python-dotenv
flask[async]